use float32 for intermediate array

+++ indi_allsky/image.py
@@ -3168,11 +3168,11 @@ class ImageProcessor(object):
 
         if self.max_bit_depth == 8:
             data_max = 256
-            range_array = numpy.arange(0, data_max).astype(numpy.uint8)
+            range_array = numpy.arange(0, data_max, dtype=numpy.float32)
             lut = (((range_array / data_max) ** (1 / float(gamma))) * data_max).astype(numpy.uint8)
         else:
             data_max = 2 ** self.max_bit_depth
-            range_array = numpy.arange(0, data_max).astype(numpy.uint16)
+            range_array = numpy.arange(0, data_max, dtype=numpy.float32)
             lut = (((range_array / data_max) ** (1 / float(gamma))) * data_max).astype(numpy.uint16)
